+++ CHANGELOG.md
@@ -4,8 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Added
+- Utility class to ease close objects without throw on error.
+
 ### Changed
 - Avoid buffer overruns in the NdrBuffer class (Based on [j-interop patch 1](https://sourceforge.net/p/j-interop/patches/1/)).
+- Simplify JIComTransport (Inspired by [j-interop patch 4](https://sourceforge.net/p/j-interop/patches/4/)).
 - Reset `JIComServer` binding to `null` when unused (Based on [j-interop patch 3](https://sourceforge.net/p/j-interop/patches/3/)).
 - Start `JIComOxidRuntime` ping timer threads on demand (Based on [j-interop patch 3](https://sourceforge.net/p/j-interop/patches/3/)).